Add Wayland support
The solution for Wayland depends on DBus. It was not ideal to
force devs to compile with the `dbus` feature if they wanted Wayland
support, because that would leave X11 users without GNOME with a lesser
experience than they have now.

To solve this, the two Linux implementations have been consolidated.